# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iTaxCix API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
...
        "/admin/role/create": {
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "Crear rol",
                "description": "Crea un nuevo rol en el sistema.",
                "operationId": "createRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleCreateRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Creado"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Petición inválida"
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Campos requeridos faltantes"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Token inválido o expirado"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/role/update": {
            "put": {
                "tags": [
                    "Role"
                ],
                "summary": "Actualizar rol",
                "description": "Actualiza un rol existente.",
                "operationId": "updateRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleUpdateRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/role/delete": {
            "delete": {
                "tags": [
                    "Role"
                ],
                "summary": "Eliminar rol",
                "description": "Elimina un rol del sistema.",
                "operationId": "deleteRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleDeleteRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Rol eliminado exitosamente (sin contenido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/role/list": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Listar roles",
                "description": "Lista todos los roles del sistema.",
                "operationId": "listRoles",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/RoleResponseDTO"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "perPage": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "currentPage": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "lastPage": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoleCreateRequestDTO": {
                "description": "Datos para crear un rol",
                "properties": {
                    "name": {
                        "description": "Nombre del rol",
                        "type": "string",
                        "example": "Administrador"
                    },
                    "active": {
                        "description": "Si el rol está activo",
                        "type": "boolean",
                        "example": true
                    },
                    "web": {
                        "description": "Si el rol es para web",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "RoleDeleteRequestDTO": {
                "description": "Datos para eliminar un rol",
                "properties": {
                    "id": {
                        "description": "ID del rol",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "RoleResponseDTO": {
                "description": "Datos de respuesta de rol",
                "properties": {
                    "id": {
                        "description": "ID del rol",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del rol",
                        "type": "string",
                        "example": "Administrador"
                    },
                    "active": {
                        "description": "Si el rol está activo",
                        "type": "boolean",
                        "example": true
                    },
                    "web": {
                        "description": "Si el rol es para web",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "RoleUpdateRequestDTO": {
                "description": "Datos para actualizar un rol",
                "properties": {
                    "id": {
                        "description": "ID del rol",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del rol",
                        "type": "string",
                        "example": "Administrador"
                    },
                    "active": {
                        "description": "Si el rol está activo",
                        "type": "boolean",
                        "example": true
                    },
                    "web": {
                        "description": "Si el rol es para web",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}